Added task view UI
Modified the UI to display detailed task information on the right.
The TreeView now only shows the task names. Clicking on a task shows
the rest of the information for the task in a detailed side view.

This view's controls are disabled once the task is complete, but
sub-tasks can still be created.

There are still some things to tweak, but this is a start. I think
it looks a lot better.