Removed task tree view, replaced with list view, much more
Replaced the task tree view with a list view. As part of this, I
added another layer to the Model/View system. Rather than the model
having data about the project, there's a new project class which
then has the task list.

I don't have any unit tests, so this is going to require some more
manual testing, but as far as I can tell, most major functionality
is working as expected.